+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -36,6 +36,7 @@ however, it has to be formatted properly to pass verification tests.
 
 - `InputUser` in combination with touchscreens no longer throws `InvalidOperationException` complaining about incorrect state format.
  * In a related change, `InputControlExtensions.GetStatePtrFromStateEvent` now works with touch events, too.
+- Stack overflow in `InputTestFixture.currentTime` getter.
 
 ## [1.0.0-preview.3] - 2019-11-14